requirements.txt
streamlit>=1.22.0
pandas>=1.3.5
numpy>=1.20.0
requests>=2.28.0
python-dotenv>=0.20.0
plotly>=5.10.0
folium>=0.12.0
streamlit-folium>=0.7.0
openai>=0.27.0
google-generativeai>=0.1.0
matplotlib>=3.5.0
seaborn>=0.12.0
pydeck>=0.8.0
openpyxl>=3.0.10
fpdf>=1.7.2 